+++ apps/web/src/pages/Schedule.test.tsx
@@ -69,6 +69,8 @@ const stub = (
   postComment: () => Promise.reject(new Error('postComment is not stubbed here')),
   updateComment: () => Promise.reject(new Error('updateComment is not stubbed here')),
   deleteComment: () => Promise.reject(new Error('deleteComment is not stubbed here')),
+  supportComment: () => Promise.reject(new Error('supportComment is not stubbed here')),
+  withdrawSupportForComment: () => Promise.reject(new Error('withdrawSupportForComment is not stubbed here')),
   offerSession: () => Promise.reject(new Error('offerSession is not stubbed here')),
   getEventAttendees: () =>
     Promise.resolve({